+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/bean/entity/HibernateUser.java
@@ -14,8 +14,8 @@ import java.util.Set;
 @Table(name = "tbl_user")
 public class HibernateUser implements Bean {
 
-    private static final long serialVersionUID = -6763701875229339369L;
-    
+    private static final long serialVersionUID = -2937876633185483767L;
+
     // -----------------------------------------------------------主键-----------------------------------------------------------
     @Id
     @Column(name = "id", length = Constraints.LENGTH_ID, nullable = false, unique = true)
@@ -31,9 +31,6 @@ public class HibernateUser implements Bean {
     // -----------------------------------------------------------一对多-----------------------------------------------------------
     @OneToMany(cascade = CascadeType.MERGE, targetEntity = HibernateMeta.class, mappedBy = "user")
     private Set<HibernateMeta> metas = new HashSet<>();
-
-    @OneToMany(cascade = CascadeType.MERGE, targetEntity = HibernateSendHistory.class, mappedBy = "user")
-    private Set<HibernateSendHistory> sendHistories = new HashSet<>();
 
     @OneToMany(cascade = CascadeType.MERGE, targetEntity = HibernateNotifySendRecord.class, mappedBy = "user")
     private Set<HibernateNotifySendRecord> sendRecords = new HashSet<>();
@@ -81,14 +78,6 @@ public class HibernateUser implements Bean {
 
     public void setMetas(Set<HibernateMeta> metas) {
         this.metas = metas;
-    }
-
-    public Set<HibernateSendHistory> getSendHistories() {
-        return sendHistories;
-    }
-
-    public void setSendHistories(Set<HibernateSendHistory> sendHistories) {
-        this.sendHistories = sendHistories;
     }
 
     public Set<HibernateNotifySendRecord> getSendRecords() {